requirements.txt
paddleocr
paddlepaddle
pillow
opencv-python
opencv-contrib-python
numpy
pytest
scikit-image
shapely
pyclipper
lmdb
tqdm
rapidfuzz
cython
pyyaml
requests
albumentations
albucore
packaging
